+++ packages/lodash-pro/scripts/tsup.config.ts
@@ -13,16 +13,21 @@ export const esm: Options = {
   target: 'es2020',
   outDir: 'dist/es',
   outExtension: () => ({ js: '.js' }),
-  plugins: [bundleless({ ext: '.js' })],
+  plugins: [
+    bundleless({
+      ext: '.js',
+    }),
+  ],
 }
 
 export const cjs: Options = {
   ...tsupConfig,
   format: ['cjs'],
+  dts: false,
   outDir: 'dist/lib',
   target: 'es2020',
   platform: 'node',
   noExternal: ['lodash-es', 'p-is-promise'],
   outExtension: () => ({ js: '.cjs' }),
-  plugins: [bundleless({ ext: '.cjs', bundle: true })],
+  plugins: [bundleless({ ext: '.cjs' })],
 }