+++ packages/vite-config/plugin-client.d.ts
@@ -1,1 +1,7 @@
-/// <reference types="vite-plugin-svgr/client" />
+declare module '*.svg?react' {
+  import type * as React from 'react'
+
+  const ReactComponent: React.FunctionComponent<React.ComponentProps<'svg'> & { title?: string }>
+
+  export default ReactComponent
+}